(18/04) Correção da P1 e criação do front-end com vite
