+++ front-end/src/App.jsx
@@ -1,0 +1,24 @@
+import './App.css'
+import {BrowserRouter, Route, Routes} from 'react-router-dom'
+import Home from './pages/Home'
+import Login from './pages/Login'
+
+
+function App() {
+  
+
+  return (
+    <div className="App">
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Home />}/>
+          <Route path="/Login" element={<Login />}/> 
+        </Routes>
+      
+      </BrowserRouter>
+      
+    </div>
+  )
+}
+
+export default App
