+++ front-end/src/pages/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+export default function Home(){
+    return (
+        <>
+            <h1>Página Inicial</h1>
+            <p>Ir Para a <Link to="/login">página de login</Link></p>
+        </>
+    )
+}
