--- front-end/src/pages/Login.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-export default function Login(){
-    
-    return(
-    <>
-        <h1>Página de login</h1>
-        <p>Esta será a página de login.</p>
-
-    </>
-    )
-}
