+++ front-end/src/pages/Login.jsx
@@ -1,0 +1,11 @@
+import React from  'react'
+
+
+export default function Login(){
+    return(
+        <>
+            <h1>Página de login</h1>
+            <p>Futura página de login</p>
+        </>
+    )
+}
